app/algorithms/requirements.txt
opencv-python
torch
torchvision
numpy
scikit-image
omegaconf
pytorch-lightning
kornia
pyyaml
simple-lama-inpainting
Pillow
python-dotenv
